# lib/swagger-ui/api-docs.json
{
    "swagger": "2.0",
    "info": {
        "version": "1.0.0",
        "title": "FundooHR Backend API's Specification",
        "description": "FundooHR is not a Software product, it is a thinking which make your task easy and efficient. And we are plotting our thining in software product. Which will make our work easy. Basically we are planning to make product based on some different technologies.",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "Bridgelabz.com",
            "email": "admin@bridgelabz.com"
        }
    },
    "basePath": "/",
...
    "schemes": [
        "http",
        "https"
    ],
...
    "consumes": [
        "application/json"
    ],
    "produces": [
        "application/json",
        "application/xml"
    ],
    "paths": {
        "/login": {
            "post": {
                "tags": [
                    "Login & Signup"
                ],
                "description": "It authenticates the user email from company database and it returns success response with email and token. For every REST call you need to pass token and crossponding email with some extra request body which is based on REST call's.",
                "operationId": "login",
                "produces": [
                    "application/json"
                ],
                "parameters": [{
                        "name": "username",
                        "username": "Logging username/email address",
                        "in": "formData",
                        "description": "Provide a valid email address to authenticate. e.g. your@some-domain.com",
                        "required": true,
                        "type": "string",
                        "default": "admin@bridgelabz.com"
                    },
                    {
                        "name": "password",
                        "in": "formData",
                        "description": "Provide a valid Password",
                        "required": true,
                        "type": "string",
                        "format": "password",
                        "default": "Bridge@123"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully authentcation the user",
                        "schema": {
                            "$ref": "#/definitions/loginResponseSuccess"
                        }
                    },
                    "401": {
                        "description": "Failed to authentcation user",
                        "schema": {
                            "$ref": "#/definitions/loginResponseFailure"
                        }
                    }
                }
            }
        },
        "/signup": {
            "post": {
                "tags": [
                    "Login & Signup"
                ],
                "description": "Signup the employee from company database and it returns success response.",
                "operationId": "signup",
                "produces": [
                    "application/json"
                ],
                "parameters": [{
                        "name": "username",
                        "username": "Signup username/email address",
                        "in": "formData",
                        "description": "Provide a valid email address to authenticate. e.g. your@some-domain.com",
                        "required": true,
                        "type": "string",
                        "default": "admin@bridgelabz.com"
                    },
                    {
                        "name": "password",
                        "in": "formData",
                        "description": "Provide a valid Password",
                        "required": true,
                        "type": "string",
                        "format": "password",
                        "default": "Bridge@123"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully authentcation the user",
                        "schema": {
                            "$ref": "#/definitions/loginResponseSuccess"
                        }
                    },
                    "401": {
                        "description": "Failed to authentcation user",
                        "schema": {
                            "$ref": "#/definitions/loginResponseFailure"
                        }
                    }
                }
            }
        },
...
        "/readEmployeeMonthAttendance": {
            "get": {
                "tags": [
                    "Attendance",
                    "Employee"
                ],
                "description": "Shows the month attendance record of an employee.",
                "operationId": "employeeMonthAttendance",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "employeeID",
                        "in": "query",
                        "description": "Employee ID is unique for each employee",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "date",
                        "date": "Date to extract month",
                        "in": "query",
                        "description": "Date parameter to get the Month & Year. Format needs to be in yyyy/mm/dd for .e.g. 2017/01/01",
                        "required": true,
                        "type": "integer"
                    },
                    {
                        "name": "x-access-token",
                        "in": "header",
                        "description": "Token needs to validate & authenticate",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success in getting employee attendance record for a month",
                        "schema": {
                            "$ref": "#/definitions/employeeMonthAttendanceSuccess"
                        }
                    },
                    "401": {
                        "description": "Failed to get the employee attendance record for a month",
                        "schema": {
                            "$ref": "#/definitions/employeeMonthAttendanceFailure"
                        }
                    }
                }
            }
        },
        "/createEmployeeDayAttendance": {
            "post": {
                "tags": [
                    "Attendance",
                    "Employee"
                ],
                "description": "Create the employee attendance record",
                "operationId": "recordEmployeeDayAttendance",
                "produces": [
                    "application/x-www-form-urlencoded"
                ],
                "parameters": [
                    {
                        "name": "Scehma",
                        "in": "body",
                        "description": "JSON format for sending the employee details & attendance details.",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/recordEmployeeDayAttendanceSchema"
                        }
                    },
                    {
                        "name": "x-access-token",
                        "in": "header",
                        "description": "Token needs to validate & authenticate",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success in getting employee attendance record for a month",
                        "schema": {
                            "$ref": "#/definitions/recordEmployeeDayAttendanceSuccess"
                        }
                    },
                    "401": {
                        "description": "Failed to get the employee attendance record for a month",
                        "schema": {
                            "$ref": "#/definitions/recordEmployeeDayAttendanceFailure"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "loginResponseSuccess": {
            "properties": {
                "status": {
                    "required": true,
                    "type": "Interger",
                    "example": true
                },
                "message": {
                    "required": true,
                    "type": "String",
                    "example": "User logged in login Successfully"
                },
                "token": {
                    "required": true,
                    "type": "String",
                    "example": "eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9.eyJzdWIiOiJhZG1pbkBicmlkZ2VsYWJ6LmNvbSIsImlhdCI6MTQ4Nzc2MDY3NSwiZXhwIjoxNDg4OTcwMjc1fQ.ALOhIMuqtczsrZkR-vtTbceP5djAQNk7CCwoZzKkOCY"
                }
            }
        },
        "loginResponseFailure": {
            "properties": {
                "status": {
                    "required": true,
                    "type": "Interger",
                    "example": false
                },
                "message": {
                    "required": true,
                    "type": "String",
                    "example": "Failed to authenticate user"
                }
            }
        },
...
        "recordEmployeeDayAttendanceSchema": {
            "properties": {
                "attendanceDate": {
                    "type": "string",
                    "required": true,
                    "example": "2017/04/01"
                },
                "employeeID": {
                    "type": "string",
                    "required": true,
                    "example": "40001EI"
                },
                "isPresent": {
                    "type": "Boolean",
                    "required": true,
                    "example": true
                },
                "inTime": {
                    "type": "Date",
                    "required": true,
                    "example": "8:30 AM"
                },
                "outTime": {
                    "type": "Date",
                    "required": true,
                    "example": "7:30 PM"
                },
                "reason": {
                    "type": "string",
                    "required": true,
                    "example": "-"
                }
            }
        },
        "recordEmployeeDayAttendanceSuccess": {
            "properties": {
                "status": {
                    "required": true,
                    "type": "Interger",
                    "example": true
                },
                "message": {
                    "required": true,
                    "type": "String",
                    "example": "Successfully created employee day attendance record"
                },
                "employeeID": {
                    "required": true,
                    "type": "string",
                    "example": "40001EI"
                },
                "date": {
                    "required": true,
                    "type": "string",
                    "example": "01/04/2017"
                }
            }
        },
        "recordEmployeeDayAttendanceFailure": {
            "properties": {
                "status": {
                    "required": true,
                    "type": "Interger",
                    "example": false
                },
                "message": {
                    "required": true,
                    "type": "String",
                    "example": "Failed to create employee day attendance record"
                }
            }
        },
        "employeeMonthAttendanceSuccess": {
            "properties": {
                "status": {
                    "required": true,
                    "type": "Interger",
                    "example": true
                },
                "message": {
                    "required": true,
                    "type": "String",
                    "example": "Successfully retrived the employee month records"
                },
                "attendanceData": {
                    "required": true,
                    "type": "object",
                    "example": [{
                            "attendanceDate":"2017/04/1",
                            "isPresent": true,
                            "intime": "8:30 AM",
                            "outtime": "2:30 PM",
                            "reason": "Leave early."
                        },{
                          "attendanceDate":"2017/04/2",
                          "isPresent": true,
                            "intime": "-",
                            "outtime": "-",
                            "reason": "My Wish"
                        }],
                "employeeData": {
                    "required": true,
                    "type": "object",
                    "example": {
                        "name": "Amit Sharma",
                        "status": "Fellowship",
                        "company": "BridgeLabz",
                        "mobile": "9594368083",
                        "emailAddress": "xyz@gmail.com",
                        "BLStartDate": "1477308028",
                        "companyJoinDate": "1477308028",
                        "companyLeaveDate": "1477308028",
                        "leaveTaken": 7,
                        "imageUrl": "http://192.168.1.102:3000/image2.jpg"
                    }
                }
            }
          }
        },
        "employeeMonthAttendanceFailure": {
            "properties": {
                "status": {
                    "required": true,
                    "type": "Interger",
                    "example": false
                },
                "message": {
                    "required": true,
                    "type": "String",
                    "example": "Failed to retrived the employee month records"
                }
            }
        },
...
    }
}